requirements.txt
streamlit
openai
azure-cognitiveservices-speech
python-dotenv
reportlab
pillow